requirements.txt
certifi==2022.12.7
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.6.1
et-xmlfile==1.1.0
lxml==4.9.2
numpy==1.24.1
openpyxl==3.1.0
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
urllib3==1.26.14
